feat: make cache generic with specific hazelcast implementation
